requirements.txt
docx2pdf>=0.1.8
python-docx>=1.1.0
PyPDF2>=3.0.0
pymupdf>=1.23.0
Pillow>=10.0.0







